@@ -1,8 +1,11 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import App from "./App";
-import * as serviceWorker from "./serviceWorker";
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import App from './App';
 
-ReactDOM.render(<App />, document.getElementById("root"));
+const root = ReactDOM.createRoot(document.getElementById('root'));
 
-serviceWorker.unregister();
+root.render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>
+);
