@@ -1,27 +1,27 @@
-import React, { Component, Fragment } from "react";
-import "bootstrap/dist/css/bootstrap.min.css";
-import "./css/styles.css";
+import React, { Component, Fragment } from 'react';
+import 'bootstrap/dist/css/bootstrap.min.css';
+import './css/styles.css';
 
-import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
+import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 
-// Global Compoennts
-import Header from "./components/Header/Header";
-import Footer from "./components/Footer/Footer";
+// Global Components
+import Header from './components/Header/Header';
+import Footer from './components/Footer/Footer';
 
 //Error Page
-import ErrorPage from "./components/Error/Error";
+import ErrorPage from './components/Error/Error';
 
 //CartList
-import CartList from "./components/CartList/CartList";
+import CartList from './components/CartList/CartList';
 
 // PDP PLP
-import ProductList from "./components/ProductList/ProductList";
+import ProductList from './components/ProductList/ProductList';
 
 //ProductDetail
-import ProductDetail from "./components/ProductDetail/ProductDetail";
+import ProductDetail from './components/ProductDetail/ProductDetail';
 
 //Edit page
-import Edit from "./components/Editor/Edit";
+import Edit from './components/Editor/Edit';
 
 class App extends Component {
   render() {
